feat(backend): moved file tagging feature to EE (#1981)
