+++ backend/pkg/objectstorage/s3/tags.go
@@ -1,0 +1,5 @@
+package s3
+
+func tagging(useTags bool) *string {
+	return nil
+}
